[I] - Provide thoughtful tests and try new definitions on own project.
