+++ types/zoomist/tslint.json
@@ -1,1 +1,3 @@
-{ "extends": "dtslint/dt.json" }
+{
+    "extends": "@definitelytyped/dtslint/dt.json"
+}
